Low-hanging bugs, unit test and code restructure
