+++ Makefile
@@ -23,7 +23,7 @@ fmt:
 
 # `make test` to run unit tests
 test: gopath
-	GOPATH=$(GOPATH) GO15VENDOREXPERIMENT=1 go test github.com/aichaos/rivescript-go/src
+	GOPATH=$(GOPATH) GO15VENDOREXPERIMENT=1 go test ./src
 
 # `make build` to build the binary
 build: gopath
