+++ doc_test.go
@@ -1,22 +1,40 @@
-package rivescript_test
+package rivescript
 
 import (
 	"fmt"
 
 	"github.com/aichaos/rivescript-go"
-	"github.com/aichaos/rivescript-go/config"
 	"github.com/aichaos/rivescript-go/lang/javascript"
 	rss "github.com/aichaos/rivescript-go/src"
 )
 
-func ExampleRiveScript() {
-	bot := rivescript.New(config.Basic())
+func Example() {
+	// Create a new RiveScript instance, which represents an individual bot
+	// with its own brain and memory of users.
+	//
+	// You can provide a rivescript.Config struct to configure the bot and
+	// provide values that differ from the defaults:
+	bot := rivescript.New(&rivescript.Config{
+		UTF8:  true, // enable UTF-8 mode
+		Debug: true, // enable debug mode
+	})
+
+	// Or if you want the default configuration, provide a nil config.
+	// See the documentation for the rivescript.Config type for information
+	// on what the defaults are.
+	bot = rivescript.New(nil)
 
 	// Load a directory full of RiveScript documents (.rive files)
 	bot.LoadDirectory("eg/brain")
 
 	// Load an individual file.
 	bot.LoadFile("testsuite.rive")
+
+	// Stream in more RiveScript code dynamically from a string.
+	bot.Stream(`
+		+ hello bot
+		- Hello, human!
+	`)
 
 	// Sort the replies after loading them!
 	bot.SortReplies()
@@ -26,10 +44,30 @@ func ExampleRiveScript() {
 	fmt.Printf("The bot says: %s", reply)
 }
 
+func ExampleRiveScript_utf8() {
+	// Examples of using UTF-8 mode in RiveScript.
+	bot := rivescript.New(rivescript.WithUTF8())
+
+	bot.Stream(`
+		// Without UTF-8 mode enabled, this trigger would be a syntax error
+		// for containing non-ASCII characters; but in UTF-8 mode you can use it.
+		+ comment ça va
+		- ça va bien.
+	`)
+
+	// Always call SortReplies when you're done loading replies.
+	bot.SortReplies()
+
+	// Without UTF-8 mode enabled, the user's message "comment ça va" would
+	// have the ç symbol removed; but in UTF-8 mode it's preserved and can
+	// match the trigger we defined.
+	reply := bot.Reply("local-user", "Comment ça va?")
+	fmt.Println(reply) // "ça va bien."
+}
+
 func ExampleRiveScript_javascript() {
 	// Example for configuring the JavaScript object macro handler via Otto.
-
-	bot := rivescript.New(config.Basic())
+	bot := rivescript.New(nil)
 
 	// Create the JS handler.
 	bot.SetHandler("javascript", javascript.New(bot))
@@ -66,8 +104,7 @@ func ExampleRiveScript_javascript() {
 func ExampleRiveScript_subroutine() {
 	// Example for defining a Go function as an object macro.
 	// import rss "github.com/aichaos/rivescript-go/src"
-
-	bot := rivescript.New(config.Basic())
+	bot := rivescript.New(nil)
 
 	// Define an object macro named `setname`
 	bot.SetSubroutine("setname", func(rs *rss.RiveScript, args []string) string {
