+++ rivescript.go
@@ -13,89 +13,65 @@ package rivescript
 */
 
 import (
-	"fmt"
-	"os"
-
-	"github.com/aichaos/rivescript-go/config"
 	"github.com/aichaos/rivescript-go/macro"
 	"github.com/aichaos/rivescript-go/sessions"
+	"github.com/aichaos/rivescript-go/sessions/memory"
 	"github.com/aichaos/rivescript-go/src"
 )
 
-const VERSION string = "0.1.0"
+// VERSION describes the module version.
+const VERSION string = "0.1.1"
 
+// RiveScript represents an individual chatbot instance.
 type RiveScript struct {
 	rs *rivescript.RiveScript
 }
 
-func New(config *config.Config) *RiveScript {
-	bot := new(RiveScript)
-	bot.rs = rivescript.New(config)
+/*
+New creates a new RiveScript instance.
+
+A RiveScript instance represents one chat bot personality; it has its own
+replies and its own memory of user data. You could make multiple bots in the
+same program, each with its own replies loaded from different sources.
+*/
+func New(cfg *Config) *RiveScript {
+	bot := &RiveScript{
+		rs: rivescript.New(),
+	}
+
+	// If no config was given, default to the BasicConfig.
+	if cfg == nil {
+		cfg = &Config{
+			Strict: true,
+			Depth:  50,
+		}
+	}
+
+	// If no session manager configured, default to the in-memory one.
+	if cfg.SessionManager == nil {
+		cfg.SessionManager = memory.New()
+	}
+
+	// Default depth if not given is 50.
+	if cfg.Depth <= 0 {
+		cfg.Depth = 50
+	}
+
+	bot.rs.Configure(cfg.Debug, cfg.Strict, cfg.UTF8, cfg.Depth, memory.New())
+
 	return bot
 }
 
-func deprecated(name string) {
-	fmt.Fprintf(os.Stderr, "Use of 'rivescript.%s()' is deprecated\n", name)
-}
-
 // Version returns the RiveScript library version.
-func (self *RiveScript) Version() string {
+func (rs *RiveScript) Version() string {
 	return VERSION
-}
-
-// SetDebug enables or disable debug mode.
-func (self *RiveScript) SetDebug(value bool) {
-	deprecated("SetDebug")
-	self.rs.Debug = value
-}
-
-// GetDebug tells you the current status of the debug mode.
-func (self *RiveScript) GetDebug() bool {
-	deprecated("GetDebug")
-	return self.rs.Debug
-}
-
-// SetUTF8 enables or disabled UTF-8 mode.
-func (self *RiveScript) SetUTF8(value bool) {
-	deprecated("SetUTF8")
-	self.rs.UTF8 = value
-}
-
-// GetUTF8 returns the current status of UTF-8 mode.
-func (self *RiveScript) GetUTF8() bool {
-	deprecated("GetUTF8")
-	return self.rs.UTF8
 }
 
 // SetUnicodePunctuation allows you to override the text of the unicode
 // punctuation regexp. Provide a string literal that will validate in
 // `regexp.MustCompile()`
-func (self *RiveScript) SetUnicodePunctuation(value string) {
-	self.rs.SetUnicodePunctuation(value)
-}
-
-// SetDepth lets you override the recursion depth limit (default 50).
-func (self *RiveScript) SetDepth(value uint) {
-	deprecated("SetDepth")
-	self.rs.Depth = value
-}
-
-// GetDepth returns the current recursion depth limit.
-func (self *RiveScript) GetDepth() uint {
-	deprecated("GetDepth")
-	return self.rs.Depth
-}
-
-// SetStrict enables strict syntax checking when parsing RiveScript code.
-func (self *RiveScript) SetStrict(value bool) {
-	deprecated("SetStrict")
-	self.rs.Strict = value
-}
-
-// GetStrict returns the strict syntax check setting.
-func (self *RiveScript) GetStrict() bool {
-	deprecated("GetStrict")
-	return self.rs.Strict
+func (rs *RiveScript) SetUnicodePunctuation(value string) {
+	rs.rs.SetUnicodePunctuation(value)
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -109,8 +85,8 @@ Parameters
 
 	path: Path to a RiveScript source file.
 */
-func (self *RiveScript) LoadFile(path string) error {
-	return self.rs.LoadFile(path)
+func (rs *RiveScript) LoadFile(path string) error {
+	return rs.rs.LoadFile(path)
 }
 
 /*
@@ -122,8 +98,8 @@ Parameters
 	extensions...: List of file extensions to filter on, default is
 	               '.rive' and '.rs'
 */
-func (self *RiveScript) LoadDirectory(path string, extensions ...string) error {
-	return self.rs.LoadDirectory(path, extensions...)
+func (rs *RiveScript) LoadDirectory(path string, extensions ...string) error {
+	return rs.rs.LoadDirectory(path, extensions...)
 }
 
 /*
@@ -134,8 +110,8 @@ Parameters
 	code: Raw source code of a RiveScript document, with line breaks after
 	      each line.
 */
-func (self *RiveScript) Stream(code string) error {
-	return self.rs.Stream(code)
+func (rs *RiveScript) Stream(code string) error {
+	return rs.rs.Stream(code)
 }
 
 /*
@@ -145,8 +121,8 @@ After you have finished loading your RiveScript code, call this method to
 populate the various sort buffers. This is absolutely necessary for reply
 matching to work efficiently!
 */
-func (self *RiveScript) SortReplies() {
-	self.rs.SortReplies()
+func (rs *RiveScript) SortReplies() {
+	rs.rs.SortReplies()
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -161,8 +137,8 @@ Parameters
 	lang: What your programming language is called, e.g. "javascript"
 	handler: An implementation of macro.MacroInterface.
 */
-func (self *RiveScript) SetHandler(name string, handler macro.MacroInterface) {
-	self.rs.SetHandler(name, handler)
+func (rs *RiveScript) SetHandler(name string, handler macro.MacroInterface) {
+	rs.rs.SetHandler(name, handler)
 }
 
 /*
@@ -172,8 +148,8 @@ Parameters
 
 	lang: The programming language for the handler to remove.
 */
-func (self *RiveScript) RemoveHandler(lang string) {
-	self.rs.RemoveHandler(lang)
+func (rs *RiveScript) RemoveHandler(lang string) {
+	rs.rs.RemoveHandler(lang)
 }
 
 /*
@@ -184,8 +160,8 @@ Parameters
 	name: The name of your subroutine for the `<call>` tag in RiveScript.
 	fn: A function with a prototype `func(*RiveScript, []string) string`
 */
-func (self *RiveScript) SetSubroutine(name string, fn rivescript.Subroutine) {
-	self.rs.SetSubroutine(name, fn)
+func (rs *RiveScript) SetSubroutine(name string, fn rivescript.Subroutine) {
+	rs.rs.SetSubroutine(name, fn)
 }
 
 /*
@@ -195,8 +171,8 @@ Parameters
 
 	name: The name of the object macro to be deleted.
 */
-func (self *RiveScript) DeleteSubroutine(name string) {
-	self.rs.DeleteSubroutine(name)
+func (rs *RiveScript) DeleteSubroutine(name string) {
+	rs.rs.DeleteSubroutine(name)
 }
 
 /*
@@ -205,8 +181,8 @@ SetGlobal sets a global variable.
 This is equivalent to `! global` in RiveScript. Set the value to `undefined`
 to delete a global.
 */
-func (self *RiveScript) SetGlobal(name, value string) {
-	self.rs.SetGlobal(name, value)
+func (rs *RiveScript) SetGlobal(name, value string) {
+	rs.rs.SetGlobal(name, value)
 }
 
 /*
@@ -215,8 +191,8 @@ GetGlobal gets a global variable.
 This is equivalent to `<env name>` in RiveScript. Returns `undefined` if the
 variable isn't defined.
 */
-func (self *RiveScript) GetGlobal(name string) (string, error) {
-	return self.rs.GetGlobal(name)
+func (rs *RiveScript) GetGlobal(name string) (string, error) {
+	return rs.rs.GetGlobal(name)
 }
 
 /*
@@ -225,8 +201,8 @@ SetVariable sets a bot variable.
 This is equivalent to `! var` in RiveScript. Set the value to `undefined`
 to delete a bot variable.
 */
-func (self *RiveScript) SetVariable(name, value string) {
-	self.rs.SetVariable(name, value)
+func (rs *RiveScript) SetVariable(name, value string) {
+	rs.rs.SetVariable(name, value)
 }
 
 /*
@@ -235,8 +211,8 @@ GetVariable gets a bot variable.
 This is equivalent to `<bot name>` in RiveScript. Returns `undefined` if the
 variable isn't defined.
 */
-func (self *RiveScript) GetVariable(name string) (string, error) {
-	return self.rs.GetVariable(name)
+func (rs *RiveScript) GetVariable(name string) (string, error) {
+	return rs.rs.GetVariable(name)
 }
 
 /*
@@ -245,8 +221,8 @@ SetSubstitution sets a substitution pattern.
 This is equivalent to `! sub` in RiveScript. Set the value to `undefined`
 to delete a substitution.
 */
-func (self *RiveScript) SetSubstitution(name, value string) {
-	self.rs.SetSubstitution(name, value)
+func (rs *RiveScript) SetSubstitution(name, value string) {
+	rs.rs.SetSubstitution(name, value)
 }
 
 /*
@@ -255,8 +231,8 @@ SetPerson sets a person substitution pattern.
 This is equivalent to `! person` in RiveScript. Set the value to `undefined`
 to delete a person substitution.
 */
-func (self *RiveScript) SetPerson(name, value string) {
-	self.rs.SetPerson(name, value)
+func (rs *RiveScript) SetPerson(name, value string) {
+	rs.rs.SetPerson(name, value)
 }
 
 /*
@@ -265,8 +241,8 @@ SetUservar sets a variable for a user.
 This is equivalent to `<set>` in RiveScript. Set the value to `undefined`
 to delete a substitution.
 */
-func (self *RiveScript) SetUservar(username, name, value string) {
-	self.rs.SetUservar(username, name, value)
+func (rs *RiveScript) SetUservar(username, name, value string) {
+	rs.rs.SetUservar(username, name, value)
 }
 
 /*
@@ -275,8 +251,8 @@ SetUservars sets a map of variables for a user.
 Set multiple user variables by providing a map[string]string of key/value pairs.
 Equivalent to calling `SetUservar()` for each pair in the map.
 */
-func (self *RiveScript) SetUservars(username string, data map[string]string) {
-	self.rs.SetUservars(username, data)
+func (rs *RiveScript) SetUservars(username string, data map[string]string) {
+	rs.rs.SetUservars(username, data)
 }
 
 /*
@@ -285,8 +261,8 @@ GetUservar gets a user variable.
 This is equivalent to `<get name>` in RiveScript. Returns `undefined` if the
 variable isn't defined.
 */
-func (self *RiveScript) GetUservar(username, name string) (string, error) {
-	return self.rs.GetUservar(username, name)
+func (rs *RiveScript) GetUservar(username, name string) (string, error) {
+	return rs.rs.GetUservar(username, name)
 }
 
 /*
@@ -294,8 +270,8 @@ GetUservars gets all the variables for a user.
 
 This returns a `map[string]string` containing all the user's variables.
 */
-func (self *RiveScript) GetUservars(username string) (*sessions.UserData, error) {
-	return self.rs.GetUservars(username)
+func (rs *RiveScript) GetUservars(username string) (*sessions.UserData, error) {
+	return rs.rs.GetUservars(username)
 }
 
 /*
@@ -304,18 +280,18 @@ GetAllUservars gets all the variables for all the users.
 This returns a map of username (strings) to `map[string]string` of their
 variables.
 */
-func (self *RiveScript) GetAllUservars() map[string]*sessions.UserData {
-	return self.rs.GetAllUservars()
+func (rs *RiveScript) GetAllUservars() map[string]*sessions.UserData {
+	return rs.rs.GetAllUservars()
 }
 
 // ClearAllUservars clears all variables for all users.
-func (self *RiveScript) ClearAllUservars() {
-	self.rs.ClearAllUservars()
+func (rs *RiveScript) ClearAllUservars() {
+	rs.rs.ClearAllUservars()
 }
 
 // ClearUservars clears all a user's variables.
-func (self *RiveScript) ClearUservars(username string) {
-	self.rs.ClearUservars(username)
+func (rs *RiveScript) ClearUservars(username string) {
+	rs.rs.ClearUservars(username)
 }
 
 /*
@@ -324,8 +300,8 @@ FreezeUservars freezes the variable state of a user.
 This will clone and preserve the user's entire variable state, so that it
 can be restored later with `ThawUservars()`.
 */
-func (self *RiveScript) FreezeUservars(username string) error {
-	return self.rs.FreezeUservars(username)
+func (rs *RiveScript) FreezeUservars(username string) error {
+	return rs.rs.FreezeUservars(username)
 }
 
 /*
@@ -336,13 +312,13 @@ The `action` can be one of the following:
 * discard: Don't restore the variables, just delete the frozen copy.
 * keep: Keep the frozen copy after restoring.
 */
-func (self *RiveScript) ThawUservars(username string, action sessions.ThawAction) error {
-	return self.rs.ThawUservars(username, action)
+func (rs *RiveScript) ThawUservars(username string, action sessions.ThawAction) error {
+	return rs.rs.ThawUservars(username, action)
 }
 
 // LastMatch returns the user's last matched trigger.
-func (self *RiveScript) LastMatch(username string) (string, error) {
-	return self.rs.LastMatch(username)
+func (rs *RiveScript) LastMatch(username string) (string, error) {
+	return rs.rs.LastMatch(username)
 }
 
 /*
@@ -352,8 +328,8 @@ This is only useful from within an object macro, to get the ID of the user who
 invoked the macro. This value is set at the beginning of `Reply()` and unset
 at the end, so this function will return empty outside of a reply context.
 */
-func (self *RiveScript) CurrentUser() string {
-	return self.rs.CurrentUser()
+func (rs *RiveScript) CurrentUser() string {
+	return rs.rs.CurrentUser()
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -368,8 +344,8 @@ Parameters
 	username: The name of the user requesting a reply.
 	message: The user's message.
 */
-func (self *RiveScript) Reply(username, message string) string {
-	return self.rs.Reply(username, message)
+func (rs *RiveScript) Reply(username, message string) string {
+	return rs.rs.Reply(username, message)
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -377,11 +353,11 @@ func (self *RiveScript) Reply(username, message string) string {
 ////////////////////////////////////////////////////////////////////////////////
 
 // DumpSorted is a debug method which dumps the sort tree from the bot's memory.
-func (self *RiveScript) DumpSorted() {
-	self.rs.DumpSorted()
+func (rs *RiveScript) DumpSorted() {
+	rs.rs.DumpSorted()
 }
 
 // DumpTopics is a debug method which dumps the topic structure from the bot's memory.
-func (self *RiveScript) DumpTopics() {
-	self.rs.DumpTopics()
+func (rs *RiveScript) DumpTopics() {
+	rs.rs.DumpTopics()
 }
