+++ src/base_test.go
@@ -1,4 +1,4 @@
-package rivescript_test
+package rivescript
 
 // NOTE: while these test files live in the 'src' package, they import the
 // public facing API from the root rivescript-go package.
@@ -7,34 +7,38 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/aichaos/rivescript-go"
-	"github.com/aichaos/rivescript-go/config"
+	"github.com/aichaos/rivescript-go/sessions"
+	"github.com/aichaos/rivescript-go/sessions/memory"
 )
 
 type RiveScriptTest struct {
-	bot      *rivescript.RiveScript
+	bot      *RiveScript
 	t        *testing.T
 	username string
 }
 
 func NewTest(t *testing.T) *RiveScriptTest {
-	return &RiveScriptTest{
-		bot:      rivescript.New(config.Basic()),
-		t:        t,
-		username: "soandso",
-	}
+	return NewTestWithConfig(t, false, false, memory.New())
 }
 
-func NewTestWithConfig(t *testing.T, config *config.Config) *RiveScriptTest {
-	return &RiveScriptTest{
-		bot:      rivescript.New(config),
+func NewTestWithUTF8(t *testing.T) *RiveScriptTest {
+	return NewTestWithConfig(t, false, true, memory.New())
+}
+
+func NewTestWithConfig(t *testing.T, debug, utf8 bool, ses sessions.SessionManager) *RiveScriptTest {
+	test := &RiveScriptTest{
+		bot:      New(),
 		t:        t,
 		username: "soandso",
 	}
+	test.bot.Debug = debug
+	test.bot.UTF8 = utf8
+	test.bot.sessions = ses
+	return test
 }
 
 // RS exposes the underlying RiveScript API.
-func (rst *RiveScriptTest) RS() *rivescript.RiveScript {
+func (rst *RiveScriptTest) RS() *RiveScript {
 	return rst.bot
 }
